Safari-611.1.10/LayoutTests/accessibility/mac/aria-valuetext-on-native-slider-expected.txt


This tests that aria-valuetext can be used on native elements, like input=range

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS range.valueDescription is 'AXValueDescription: hello'
PASS successfullyParsed is true

TEST COMPLETE
